Added constructor that uses last error info if approriate.
When a method is called on an interface pointer and it returns a failure, the com_error the caller should throw can include information from IErrorInfo if the interface in question supports it.  The new constructor takes this interface pointer as a paramter, queries it for ErrorInfo support and, if successful, augments the com_error with the last ErrorInfo set on the calling thread.