Trim trailine newline from what() message.
FormatMessage appends '\r\n' to the error messages it produces.  This isn't how it should be; it should be up the the caller to decide whether the messages require a new line.  For example, when displaying the message in a GUI, you are very unlikely to want a new line after the message.

This change fixes it by omitting the '\r\n' from the string it returns.