Added new enum_iterator that can iterate over the items of any COM enumerator.
The details of how the items are copied into the target need some work.  The target may be a POD, in which case we must be careful to clear them, or a smart wrapper, in which case we don't need to.  It's not clear how to do this in a way that satisfies both well.