Separate the two different ways to treat daylight savings time.
The conversion functions were muddling two different ways of choosing how to treat daylight savings time for datetimes in the local timezone.  In the first, it is calculated from a date and a flag saying whether that date is local or UTC.  In the second, the caller explicitly chooses whether to assume daylight savings is in operation.

The two methods were combined using a single set of flags that tried to do woth parts at the same time along with a date argument that would not be used if the second method was in use.  This made the code very muddled and hard to understand.

This change separates out the two modes as overloads of the existing methods.