Fix date conversion to not base DST on 30 December 1899.
The code is documented as defaulting to calculating DST from the date in the object.  But it decides whether the caller wants the default by testing whether the 'as of' date is invalid.  But the default argument was datetime_t() which is the epoch and is not invalid.  This commit changes the default argument to dt_invalid so the correct behaviour is invoked.