Removed misuse of wstring's iterator.
bstr_t was initialising a wstring::reverse_iterator from a BSTR (wchar_t*), but wstring::reverse_iterator is only defined as being constructible from a wstring::iterator.  This seems to be assuming that, if iterator debugging is not in operation, the string iterator will be a naked pointer - a wrong assumption.