Removed exception specifications throwing com_error.
The com_ptr and variant classes user a forward-declaration of com_error because com_error, in turn, uses those classes.  However, they were using com_error in their exception specifications, where it is illegal to use an incomplete type.  Rather than find a way to fix the circular dependency, which could be long and tedious, I remove the exception specifications, as their use is frowned on nowadays anyway.