Fix temporary file creation to work with 64-bit program.
We were trying to create the file somewhere we shouldn't, probably C:\.  If compiled as 32-bit, Windows would silently redirect this and all would be well.  But when compiled 64-bit, the `CreateFile` would fail.  This difference is odd but this fixes both 32 and 64-bit to create the file in the current directory if `tmpnam()` prefixes its path with a backslash.  This matches the MSDN documentataion for `tmpnam()`.

See http://stackoverflow.com/q/19416696/67013.