Fix datetime_t test suite handling of DST.
The tests were using the Windows FileTimeToLocalFileTime function to generate the expected local FILETIMEs and SYSTEMTIMEs, but this function doesn't take account of DST, which caused spurious test failures for half the year.  This change fixes that by using SystemTimeToTxSpecificLocalTime instead.

In the process we extend the tests to use two dates, one early in the year, one six months later, so that failures of this kind will cause test failures *all year round*.