Reveal delayed errors during stream wrapper Write.
The wrapped C++ streambuf may be using a buffer so write operations may appear to succeed only to fail when the buffer is next written out.  This next time may not be during the call to Write; it may even be delayed till the inner stream is destroyed - long after the wrapper has gone away.  This is not ideal from the wrapper's point of view as these errors will be lost unless there is another call to Write which overflows the buffer.

Therefore, this commit changes the wrapper to flush the inner stream after every write.  This forces the errors to appear immediately and they are reported via the error code from Write.