Changed stream wrapper `Read`/`Write` error behaviour for better performance.
Returning an exact byte-count from `Read` or `Write` in the error case meant reading/writing from/to the stream one byte at a time.  In the write case this also meant flushing the stream after every byte.  This had a massive performance impact.

This commit changes the behaviour such that the returned byte count will never be more than the amount read/written but may be less.  This is more practical and we believe is still within the documented behaviour of `IStream`.